moved tbr accepting to a new method of the LoopPlugin
