+++ app/src/main/java/info/nightscout/androidaps/events/EventAcceptOpenLoopChange.java
@@ -1,0 +1,5 @@
+package info.nightscout.androidaps.events;
+
+/** Base class for events to update the UI, mostly a specific tab. */
+public class EventAcceptOpenLoopChange extends Event {
+}
