--- app/src/main/java/info/nightscout/androidaps/plugins/Wear/events/EventWearAcceptOpenLoopChange.java
@@ -1,9 +1,0 @@
-package info.nightscout.androidaps.plugins.Wear.events;
-
-import info.nightscout.androidaps.events.Event;
-
-/**
- * Created by mike on 07.08.2016.
- */
-public class EventWearAcceptOpenLoopChange extends Event {
-}
